Move distinction between expressions and statements
Before this commit, the parser was responsible for distinguishing
between expressions and statements, and it refused to parse whole
classes of programs. Amoung the programs it refused to parse were, e.
g., `print print 5` or `5` which are valid programs, just ones that fail
with an error at runtime.

This commit changes the whole stack, from parser to VM, to be less
strict when parsing and to replace parser errors with runtime errors. It
does so by extending the compiler’s knowledge of expressions and
statements. The compiler now knows which nodes to expect in which
position and injects necessary runtime checks to make the VM fail with
an error where certain constraints are violated.