Add Logo as a central entry point
Make `Vm.run` return `State` instead of `Result Error Vm`.

The new module, `Logo`, hides the details of compiling and running
programs. In case of an error, the new `Vm.run` enables keeping the `Vm`
in the state it had before the error. This will become more important
once a UI will be added.