Extract Instruction to own module
This resolves a circular dependency.

With this commit, both `repcount` and `templateVariable` are changed so
that they now operate on `List Scope` instead of `{ a | scopes : List
Scope }`.  That makes it possible to define `Instruction` in `Vm.Vm`
without referencing the type `Vm`, thus resolving the circular
dependency `Instruction -> Vm -> Instruction`.

That, in turn, allows for extracting the type `Instruction` into its own
module which is required for making user-defined functions part of
`Environment` which is not yet possible because it would create a
circular dependency `Instruction -> Vm -> Environment -> Instruction`.