Run elm-test install elm-explorations/test
This is an attempt to make sure `elm.json` is considered valid by `elm`.

This is the beginning of the error message this commit is trying to
address:

```
It looks like the dependencies elm.json in were edited by hand (or by a 3rd
party tool) leaving them in an invalid state.
```